Updated Select Match screen, select match, it displays team numbers, highlights team corresponding to tablet ID in Alliance color. Select match, next screen shows team number in large letters with background in Alliance color.
Import uses a boolean in FTSUtilities to contol whether in test mode or
live. In Test mode, will populate data based on functions in DBAdapter
classes. In live mode, looks for csv file in files directory.

TeamMatchData saves individual match files. Import class compiles them
to a single csv file for sending to laptop, and saves the single csv
files to the 'saved' directory in case needed again. No current way to
're-export' the last export in case of failure.